# dimos/server/openapi.json
{
  "components": {
    "schemas": {
...
      "CloudLoginEvent": {
        "additionalProperties": true,
        "description": "The device login's state changed",
        "properties": {
          "login": {
            "$ref": "#/components/schemas/Login",
            "description": "The login now"
          },
          "type": {
            "const": "cloud-login",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "login"
        ],
        "title": "CloudLoginEvent",
        "type": "object",
        "x-zenoh-key": "<ns>/dimos/events/cloud-login"
      },
...
      "DimosEvent": {
        "description": "Any of the server's events; `type` says which, and is the last chunk of its zenoh key",
        "discriminator": {
          "mapping": {
            "cloud-login": "#/components/schemas/CloudLoginEvent",
            "discovery": "#/components/schemas/DiscoveryEvent",
            "job": "#/components/schemas/JobEvent",
            "launch": "#/components/schemas/LaunchEvent",
            "log": "#/components/schemas/LogEvent",
            "upload": "#/components/schemas/UploadEvent",
            "upload-removed": "#/components/schemas/UploadRemovedEvent",
            "uploads": "#/components/schemas/UploadsEvent"
          },
          "propertyName": "type"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/LaunchEvent"
          },
          {
            "$ref": "#/components/schemas/LogEvent"
          },
          {
            "$ref": "#/components/schemas/UploadEvent"
          },
          {
            "$ref": "#/components/schemas/UploadsEvent"
          },
          {
            "$ref": "#/components/schemas/UploadRemovedEvent"
          },
          {
            "$ref": "#/components/schemas/CloudLoginEvent"
          },
          {
            "$ref": "#/components/schemas/DiscoveryEvent"
          },
          {
            "$ref": "#/components/schemas/JobEvent"
          }
        ],
        "title": "DimosEvent"
      },
...
      "ErrorResponse": {
        "additionalProperties": true,
        "description": "Every error answer (4xx and 5xx).",
        "properties": {
          "error": {
            "description": "What went wrong, in words a person can act on",
            "examples": [
              "no upload u7"
            ],
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Info": {
        "additionalProperties": true,
        "description": "The dimos checkout this server launches runs from.",
        "properties": {
          "dir": {
            "description": "The checkout's folder",
            "examples": [
              "/home/me/dimos"
            ],
            "title": "Dir",
            "type": "string"
          },
          "found": {
            "description": "A dimos checkout (pyproject.toml naming dimos) is there",
            "title": "Found",
            "type": "boolean"
          },
          "inRange": {
            "description": "`version` is inside `range` (a launch is refused when not)",
            "title": "Inrange",
            "type": "boolean"
          },
          "installed": {
            "description": "Its `.venv/bin/dimos` exists, so it can run blueprints",
            "title": "Installed",
            "type": "boolean"
          },
          "range": {
            "description": "The dimos versions Desktop supports ($DESKTOP_DIMOS_RANGE); empty = any",
            "examples": [
              ">=0.0.14b1 <0.1"
            ],
            "title": "Range",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Its pyproject version (null: none found)",
            "examples": [
              "0.0.14"
            ],
            "title": "Version"
          }
        },
        "required": [
          "dir",
          "found",
          "installed",
          "version",
          "range",
          "inRange"
        ],
        "title": "Info",
        "type": "object"
      },
...
      "Ok": {
        "additionalProperties": true,
        "properties": {
          "ok": {
            "const": true,
            "description": "Always true",
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "Ok",
        "type": "object"
      },
...
      "StopRequest": {
        "additionalProperties": true,
        "properties": {
          "runId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "run id (any live run in the registry); none: the launch this server started",
            "examples": [
              "20260101-120000-unitree-go2"
            ],
            "title": "Runid"
          }
        },
        "title": "StopRequest",
        "type": "object"
      },
      "StopResult": {
        "additionalProperties": true,
        "properties": {
          "output": {
            "description": "What was stopped",
            "examples": [
              "stopped unitree-go2 (pid 41233)"
            ],
            "title": "Output",
            "type": "string"
          }
        },
        "required": [
          "output"
        ],
        "title": "StopResult",
        "type": "object"
      },
      "Stopping": {
        "additionalProperties": true,
        "properties": {
          "stopping": {
            "const": true,
            "description": "Always true: the server exits in 0.2 s",
            "title": "Stopping",
            "type": "boolean"
          }
        },
        "required": [
          "stopping"
        ],
        "title": "Stopping",
        "type": "object"
      },
...
      "UploadList": {
        "additionalProperties": true,
        "properties": {
          "uploads": {
            "description": "The queue, in order",
            "items": {
              "$ref": "#/components/schemas/Upload"
            },
            "title": "Uploads",
            "type": "array"
          },
          "waitingForLogin": {
            "description": "An upload found no cloud login: the queue waits for one",
            "title": "Waitingforlogin",
            "type": "boolean"
          }
        },
        "required": [
          "uploads",
          "waitingForLogin"
        ],
        "title": "UploadList",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The dimos server's HTTP API: blueprints, global config, runs and their logs, events, Dimensional cloud uploads,\ndiscovery (blueprints, modules, message types), docs, extras and jobs.\n\ndimOS Desktop starts the server (`python -m dimos.server`, on a unix socket) and forwards `/dimos/...` to it unchanged.\nEvery answer is JSON unless the operation says otherwise; every error is `{\"error\": \"<message>\"}` (ErrorResponse).\nState is HTTP; changes are events, published on zenoh at `<ns>/dimos/events/<type>` (the `events` tag).\n\nExtensions: `x-family: dimos` on every operation; `x-agent: true` marks what Desktop's agent finds through\nsearch_endpoints; `x-mcp-tool` names an MCP tool that does the same; `x-zenoh-key` on an event schema is its key.\n",
    "title": "dimos server",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/dimos/events": {
      "get": {
        "deprecated": true,
        "description": "Deprecated and internal, kept for one release: the same events are published on zenoh at `<ns>/dimos/events/<type>` (Desktop's docs/events.md), so listen there. The stream starts with the current `launch` event, then sends every event (a client that can't keep up loses events). Each event type's payload is a component schema (LaunchEvent, LogEvent, UploadEvent, UploadsEvent, UploadRemovedEvent, CloudLoginEvent; DimosEvent is any of them).",
        "operationId": "get_dimos_events",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/DimosEvent"
                }
              }
            },
            "description": "Server-sent events, one `data: <DimosEvent JSON>` each; `:` keep-alives every 15 s"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "The dimos server's live events (SSE): launch phases, warning+ log records, uploads, the cloud login",
        "tags": [
          "events"
        ],
        "x-agent": false,
        "x-family": "dimos"
      }
    },
...
    "/dimos/info": {
      "get": {
        "description": "Reads the checkout's pyproject.toml and whether its `.venv/bin/dimos` exists, and checks the version against the range Desktop supports ($DESKTOP_DIMOS_RANGE). A launch is refused while `inRange` is false. No side effects.",
        "operationId": "get_dimos_info",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            },
            "description": "`{ dir, found, installed, version, range, inRange }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "The dimos checkout this server drives: dir, version, installed",
        "tags": [
          "server"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
...
    "/dimos/runs/stop": {
      "post": {
        "description": "Sends the run's process group SIGINT, then SIGTERM after 20 s, then SIGKILL after 10 more, and answers once it's gone. Stops `runId` (any live run in the registry) or else this server's launch; the body is optional. 500 when there's nothing running to stop or it won't stop.",
        "operationId": "post_dimos_runs_stop",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/StopRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopResult"
                }
              }
            },
            "description": "`{ output }`"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request: a missing or malformed body or parameter, or a value the server refuses (the message says which)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Stop the blueprint this server launched (or runId)",
        "tags": [
          "runs"
        ],
        "x-agent": true,
        "x-family": "dimos",
        "x-mcp-tool": "stop_blueprint"
      }
    },
...
    "/dimos/server/stop": {
      "post": {
        "description": "Answers, then exits 0.2 s later. A running upload's worker is killed and the upload is queued again for the next start (dimos resumes it); launched blueprints keep running (their own sessions).",
        "operationId": "post_dimos_server_stop",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stopping"
                }
              }
            },
            "description": "`{ stopping: true }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Make the dimos server exit (Desktop starts it again when needed)",
        "tags": [
          "server"
        ],
        "x-agent": false,
        "x-family": "dimos"
      }
    },
    "/dimos/uploads": {
      "delete": {
        "description": "Removes every done, failed or cancelled upload from the list (what is in the cloud is still remembered, see /dimos/uploads/uploaded) and answers the queue.",
        "operationId": "delete_dimos_uploads",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadList"
                }
              }
            },
            "description": "`{ uploads: [Upload], waitingForLogin }`, without the finished ones"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Clear the finished uploads (done, failed, cancelled) from the list",
        "tags": [
          "uploads"
        ],
        "x-agent": false,
        "x-family": "dimos"
      },
      "get": {
        "description": "The queue in order, and whether it waits for a cloud login. No side effects.",
        "operationId": "get_dimos_uploads",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadList"
                }
              }
            },
            "description": "`{ uploads: [Upload], waitingForLogin }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "The Dimensional cloud upload queue: each upload's state, progress, speed, time left and error",
        "tags": [
          "uploads"
        ],
        "x-agent": true,
        "x-family": "dimos"
      },
      "post": {
        "description": "Adds the recording to the end of the queue (saved, so it survives a restart) and answers its upload, queued; one already queued or uploading for that path is answered instead. Needs a cloud login: without one it waits. 400 when the path isn't an absolute path to an existing dimos recording (an .mcap, or a .db dimos recorded).",
        "operationId": "post_dimos_uploads",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Upload"
                }
              }
            },
            "description": "`Upload`"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request: a missing or malformed body or parameter, or a value the server refuses (the message says which)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Upload a recording (.mcap or .db) to Dimensional cloud: it joins the queue (one at a time)",
        "tags": [
          "uploads"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
...
    "/dimos/uploads/{id}": {
      "delete": {
        "description": "A queued upload turns cancelled; a running one's worker is killed and it turns cancelled; a finished one is removed (an `upload-removed` event). 404 for an unknown id.",
        "operationId": "delete_dimos_uploads_id",
        "parameters": [
          {
            "description": "upload id, e.g. u3",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "upload id, e.g. u3",
              "examples": [
                "u3"
              ],
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            },
            "description": "`{ ok: true }`"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such thing (the message names it)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The server couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Cancel a queued or running upload, or remove a finished one from the list",
        "tags": [
          "uploads"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
...
  },
...
}